requirements.txt
Flask==2.2.3
openai==1.2.3
python-dotenv==1.0.0
pytest==7.2.1